dmp/requirements.txt
kafka-python==2.0.2
PyMySQL==1.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
